Test: user ドメインの振る舞いテストを user_domain_test.go へ統合（レビュー指摘②）
user_behavior_test.go（UpdateProfile / ChangePassword / MarkAsDeleted）を
ソース対応の user_domain_test.go に統合し、別ファイルを削除。
「1ソース=1テストファイル」を基準とし、肥大化時は親ソースごと分割する方針に統一。
domain/user のカバレッジは 100% を維持。
